// src/interfaces/iPayLoadDriverCommandReport.ts
// Copyright 2021 TestProject (https://testproject.io)
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//   http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

/**
 * Driver command report payload
 */
export interface IPayLoadDriverCommandReport {
  /**
   * @member {string} commandName - The command name
   */
  commandName: string;

  /**
   * @member {unkwown} commandParameters - The command parameters
   */
  commandParameters: unknown;

  /**
   * @member {unknown} result - The result
   */
  result: unknown;

  /**
   * @member {boolean} passed - True if passes otherwise false
   */
  passed: boolean;
}